+++ DrinkingPlace/MapView.swift
@@ -1,0 +1,21 @@
+//
+//  MapView.swift
+//  DrinkingPlace
+//
+//  Created by Masahiro Ichikawa on 2020/07/18.
+//  Copyright © 2020 Isobe Gundan. All rights reserved.
+//
+
+import SwiftUI
+
+struct MapView: View {
+    var body: some View {
+        Text("検索ボタンが押された")
+    }
+}
+
+struct MapView_Previews: PreviewProvider {
+    static var previews: some View {
+        MapView()
+    }
+}
